Updated command names
