+++ Command/NodeTypeListCommand.php
@@ -20,25 +20,32 @@
 
 namespace Doctrine\Bundle\PHPCRBundle\Command;
 
-use PHPCR\Util\Console\Command\ExportXmlCommand as BaseExportXmlCommand;
+use PHPCR\Util\Console\Command\NodeTypeListCommand as BaseTypeListCommand;
 
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * @author David Buchmann <david@liip.ch>
+ * @author Daniel Leech <daniel@dantleech.com>
  */
-class ExportXmlCommand extends BaseExportXmlCommand
+class NodeTypeListCommand extends BaseTypeListCommand
 {
+    /**
+     * Configures the current command.
+     */
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('doctrine:phpcr:export')
-            ->addOption('session', null, InputOption::VALUE_OPTIONAL, 'The session to use for this command')
+            ->setName('doctrine:phpcr:node-type:list')
+            ->addOption(
+                'session', null, 
+                InputOption::VALUE_OPTIONAL, 
+                'The session to use for this command'
+            )
         ;
     }
 
@@ -52,8 +59,13 @@ class ExportXmlCommand extends BaseExportXmlCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        DoctrineCommandHelper::setApplicationPHPCRSession($this->getApplication(), $input->getOption('session'));
+        DoctrineCommandHelper::setApplicationPHPCRSession(
+            $this->getApplication(), 
+            $input->getOption('session')
+        );
 
         return parent::execute($input, $output);
     }
 }
+
+
