+++ Command/NodeTypeRegisterCommand.php
@@ -20,51 +20,33 @@
 
 namespace Doctrine\Bundle\PHPCRBundle\Command;
 
-use PHPCR\Util\Console\Command\NodeTouchCommand as BaseNodeTouchCommand;
+use PHPCR\Util\Console\Command\NodeTypeRegisterCommand as BaseRegisterNodeTypesCommand;
 
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * @author Daniel Leech <daniel@dantleech.com>
+ * Wrapper to use RegisterNodeTypeCommand with Symfony's app/console
+ *
+ * @see Doctrine/ODM/PHPCR/Tools/Console/Command/RegisterNodeTypesCommand
  */
-class NodeTouchCommand extends BaseNodeTouchCommand
+class NodeTypeRegisterCommand extends BaseRegisterNodeTypesCommand
 {
-    /**
-     * Configures the current command.
-     */
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('doctrine:phpcr:node:touch')
-            ->addOption(
-                'session', null, 
-                InputOption::VALUE_OPTIONAL, 
-                'The session to use for this command'
-            )
+            ->setName('doctrine:phpcr:node-type:register')
+            ->addOption('session', null, InputOption::VALUE_OPTIONAL, 'The session to use for this command')
         ;
     }
 
-    /**
-     * Executes the current command.
-     *
-     * @param InputInterface  $input  An InputInterface instance
-     * @param OutputInterface $output An OutputInterface instance
-     *
-     * @return integer 0 if everything went fine, or an error code
-     */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        DoctrineCommandHelper::setApplicationPHPCRSession(
-            $this->getApplication(), 
-            $input->getOption('session')
-        );
+        DoctrineCommandHelper::setApplicationPHPCRSession($this->getApplication(), $input->getOption('session'));
 
         return parent::execute($input, $output);
     }
 }
-
